backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
fastapi==0.115.12
idna==3.10
mysqlclient==2.2.7
pydantic==2.11.2
pydantic_core==2.33.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
